+++ authlete/dto/client_extension.py
@@ -1,5 +1,5 @@
 #
-# Copyright (C) 2019 Authlete, Inc.
+# Copyright (C) 2019-2020 Authlete, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -22,7 +22,9 @@ class ClientExtension(Jsonable):
     def __init__(self, nameAndValues=None):
         nameAndTypes = {
             'requestableScopesEnabled': bool,
-            'requestableScopes':        str   # list of str
+            'requestableScopes':        str,   # list of str
+            'accessTokenDuration':      int,
+            'refreshTokenDuration':     int,
         }
 
         super().__init__(nameAndValues, nameAndTypes)